Modify In Place Test
What changed?
Created test to check modification of a file in place

Changed rename_and_remove_tempfile fn to take in the tuple output by mkfstemp function
instead of just the filename

Why?
In order to actually close the tempfile, we need the file decriptor. Easiest way to
get this was to use the output stored earlier from this function, as it contains the
tempfile name as well.

References